+++ app/src/main/java/com/vera/zzl/comp6442_assignment_two_2016/Operation/Multiplication.java
@@ -19,8 +19,7 @@ public class Multiplication extends Expressions {
     }
 
     @Override
-    public float evaluate(){
-        return leftPara.evaluate()*rightPara.evaluate();
+    public Float evaluate(){
+        return (float)leftPara.evaluate()*(float)rightPara.evaluate();
     }
-
 }
